heroku/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Flask==2.2.1
frozenlist==1.3.1
future==0.18.2
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.3.0
MarkupSafe==2.1.1
multidict==6.0.2
pykintone==0.3.10
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.11
Werkzeug==2.2.1
yarl==1.8.1
